test app v2

+++ putevoi/ComboItem.cs
@@ -7,10 +7,10 @@ namespace putevoi
 {
     class ComboItem
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string value { get; set; }
 
-        public ComboItem(int _id, string _text)
+        public ComboItem(long _id, string _text)
         {
             id = _id;
             value = _text;
